Added reposition function
When there isn’t enough room beneath the input the datepicker will display above the input.
When this happens and you go to the "Select Month" view, there will be a gap between the picker and the input.
This occurs because the list of months is smaller than the list of days. The same problem occurs when selecting a year.
This can be resolved by repositioning the picker.